Possible to get temperatre measurements from MCP9808
