+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.pythonPath": "venv/bin/python3.7",
+    "python.linting.enabled": true
+}
