--- icypi/requirements.txt
@@ -1,7 +1,0 @@
-bcrypt==3.1.7
-cffi==1.13.2
-cryptography==2.8
-paramiko==2.7.1
-pycparser==2.19
-PyNaCl==1.3.0
-six==1.14.0
