+++ src/get_temperature_mcp9808.py
@@ -1,0 +1,5 @@
+from icypi.sensor import Sensor
+
+sensor = Sensor('MCP9808')
+
+print(sensor.get_temperature())
